Update README.md
. (5 marks) Giangiovanni Merluscone proposes the following algorithm to do it. Merge the rst two sorted piles into a sorted pile; then merge the resulting pile with the third sorted pile; then merge the resulting pile with the fourth sorted pile, etc. What is the running time of this algorithm in terms of n and m? Recall that merging two sorted piles of papers can be done in time proportional to the size of the resulting pile.

b. (10 marks) Give a more e cient divide-and-conquer algorithm to solve this problem. Describe your algorithm clearly and concisely in English (you don’t need to give the pseudo-code). What is the (worst-case) running time of your algorithm? You don’t need to prove the correctness of your algorithm, but you must justify its running time by giving (and solving) the running time recurrence relation. For simplicity you may assume that n or m is an exact power of 2.

Question 2. (20 marks) Given an array A of n integers, some of which may be negative, we want to nd

Sij such that: